hw12: flavors free/paid. free отличается тем, что в навигации нет избранного
