+++ app/build.gradle
@@ -53,6 +53,18 @@ android {
     buildFeatures {
         viewBinding true
     }
+
+    flavorDimensions 'version'
+    productFlavors {
+        free {
+            dimension 'version'
+            versionNameSuffix '.free'
+        }
+        paid {
+            dimension 'version'
+            versionNameSuffix '.paid'
+        }
+    }
 }
 
 dependencies {
