+++ app/src/free/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:id="@+id/navigation_list"
+        android:icon="@drawable/ic_list"
+        android:title="@string/bottom_nav_list" />
+
+    <item
+        android:id="@+id/navigation_default"
+        android:icon="@drawable/ic_default"
+        android:title="@string/bottom_nav_default" />
+
+    <item
+        android:id="@+id/navigation_history"
+        android:icon="@drawable/ic_history"
+        android:title="@string/bottom_nav_history" />
+</menu>
