+++ app/src/free/res/navigation/main_navigation.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/main_navigation"
+    app:startDestination="@id/navigation_default">
+
+    <fragment
+        android:id="@+id/navigation_list"
+        android:name="com.example.weather.ui.citylist.CityListFragment"
+        android:label="@string/bottom_nav_list"
+        tools:layout="@layout/fragment_city_list" />
+
+    <fragment
+        android:id="@+id/navigation_default"
+        android:name="com.example.weather.ui.details.DetailsFragment"
+        android:label="@string/bottom_nav_default"
+        tools:layout="@layout/details_fragment" />
+
+    <fragment
+        android:id="@+id/navigation_history"
+        android:name="com.example.weather.ui.history.HistoryFragment"
+        android:label="@string/bottom_nav_history"
+        tools:layout="@layout/fragment_history" />
+</navigation>
