+++ app/src/free/res/values/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="flavor_string_value" translatable="false">Free version</string>
+</resources>
