--- app/src/main/java/com/example/weather/ui/favorites/FavoritesFragment.kt
@@ -1,9 +1,0 @@
-package com.example.weather.ui.favorites
-
-import androidx.fragment.app.Fragment
-
-class FavoritesFragment : Fragment() {
-    companion object {
-        fun newInstance() = FavoritesFragment()
-    }
-}
