+++ app/src/main/res/layout/fragment_city_list.xml
@@ -87,7 +87,12 @@
                 android:layout_height="wrap_content"
                 android:layout_weight="1"/>
             <TextView
-                android:id="@+id/version_string_text_view"
+                android:id="@+id/build_version_string_text_view"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"/>
+            <TextView
+                android:id="@+id/flavor_version_string_text_view"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_weight="1"/>
