+++ app/src/paid/java/com/example/weather/favorites/FavoritesFragment.kt
@@ -1,0 +1,9 @@
+package com.example.weather.ui.favorites
+
+import androidx.fragment.app.Fragment
+
+class FavoritesFragment : Fragment() {
+    companion object {
+        fun newInstance() = FavoritesFragment()
+    }
+}
